Model is not JIT deployable

+++ geqtrain/data/dataset.py
@@ -109,7 +109,7 @@ def parse_attrs(
                     mask = np.isnan(val)
                     if np.any(mask) and not can_be_undefined:
                         raise Exception(f"Found NaN value for attribute {key}. If this is allowed set 'can_be_undefined' to True in config file for this attribute.")
-                    _input_type = val.astype(np.int32)
+                    _input_type = val.astype(np.int64)
                     _input_type[mask] = num_types
                     # 'unkown' token has value 'num_types', while defined tokens have range [0, 'num_types')
             if key in _fields:
